backend/requirements.txt
# backend/requirements.txt
# Core API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# AWS SDKs
boto3>=1.34.0
botocore>=1.34.0

# Audio processing
pydub>=0.25.1
soundfile>=0.12.1
numpy>=1.24.0

# Optional audio helpers (safer WAV parsing)
scipy>=1.11.0

# Step Functions / workflow helpers
python-dotenv>=1.0.0
python-multipart
# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.2

# HTTP utilities
requests>=2.31.0

# Type hints
typing_extensions>=4.0.0

# Compression / parsing
orjson>=3.9.0

# For PII detection placeholder (regex-based)
regex>=2023.12.25
